Exit process with error code if the operation fails
When the license setting process failed, the process still returned with success (exit code = 0).
This made it more difficult to detect a problem: the caller process had to parse the output text of the process.

Exit code is now set to non-zero value, so the caller process can detect errors without searching for specific text patterns.